+++ lib/layrz_theme.dart
@@ -27,3 +27,4 @@ export 'src/table/table.dart';
 export 'src/utilities/utilities.dart';
 export 'src/map/map.dart';
 export 'src/colorblindness/colorblindness.dart';
+export 'src/alerts/alerts.dart';